requirements.txt
# requirements.txt

requests
python-telegram-bot==20.3
python-dotenv
python-dateutil
